fix negitive number bug

+++ config/initializers/bank_account_methods.rb
@@ -32,7 +32,7 @@ module BankAccountMethods
     class_eval <<-EOF
       def #{operation}! amount, option = {}
         ensure_positive_number(amount)
-        operate_credit!(amount.to_i, "#{operation}", option)
+        operate_credit!(amount.to_f, "#{operation}", option)
       end
     EOF
   end
@@ -60,6 +60,6 @@ module BankAccountMethods
   end
 
   def ensure_positive_number amount
-    raise NegativeNumberError.new if amount.blank? or amount.to_i <= 0
+    raise NegativeNumberError.new if amount.blank? or amount.to_f <= 0
   end
 end
